"Updated index.html title, refactored App.jsx, modified Card.jsx, and removed Gallery.jsx"
