+++ client/src/App.jsx
@@ -1,12 +1,7 @@
-import Gallery from "./pages/Gallery.jsx";
+import Gallery from "./routes/Gallery.jsx";
 function App() {
   return (
     <>
-      {/* <div className="min-h-screen flex items-center justify-center bg-gray-100">
-        <h1 className="text-3xl font-bold text-blue-500">
-          Hey there 🙋🏿‍♀️🙋🏿‍♂️ Developers 💻🖥⌨🖲, Tailwind CSS with React and Vite!
-        </h1>
-      </div> */}
       <Gallery />
     </>
   );
